+++ keyboards/gopolar/gg86/keymaps/bongocat/rules.mk
@@ -1,2 +1,4 @@
+SRC += oled/bongocat.c
+
 VIA_ENABLE = yes
 WPM_ENABLE = yes
