+++ keyboards/gopolar/gg86/rules.mk
@@ -1,3 +1,5 @@
+SRC += lib/logo.c
+
 # MCU name
 MCU = atmega32u4
